refactor(editor): style layer 파일 분리
