+++ front/src/components/editor/BlockEditorEngine.blockStyles.tsx
@@ -1,0 +1,388 @@
+import styled from "@emotion/styled"
+
+export const BlockHandleRail = styled.div`
+  position: absolute;
+  z-index: 55;
+  display: flex;
+  flex-direction: row;
+  align-items: flex-start;
+  gap: 0.18rem;
+  padding: 0.12rem;
+  border-radius: 0.72rem;
+  border: 1px solid ${({ theme }) =>
+    theme.scheme === "dark" ? "rgba(148, 163, 184, 0.2)" : "rgba(71, 85, 105, 0.14)"};
+  background: ${({ theme }) =>
+    theme.scheme === "dark" ? "rgba(15, 23, 42, 0.52)" : "rgba(255, 255, 255, 0.84)"};
+  backdrop-filter: blur(6px);
+  opacity: 0;
+  transform: translate3d(-3px, 0, 0);
+  pointer-events: none;
+  transition:
+    opacity 140ms ease,
+    transform 140ms ease;
+
+  &[data-visible="true"] {
+    opacity: 1;
+    transform: translate3d(0, 0, 0);
+    pointer-events: auto;
+  }
+`
+
+export const BlockHandleButton = styled.button`
+  all: unset;
+  box-sizing: border-box;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  width: 1.44rem;
+  height: 1.44rem;
+  border-radius: 0.42rem;
+  border: 1px solid transparent;
+  background: transparent;
+  color: ${({ theme }) => theme.colors.gray10};
+  font-size: 0.76rem;
+  font-weight: 700;
+  box-shadow: none;
+  opacity: 0.8;
+  cursor: pointer;
+  transition:
+    background-color 120ms ease,
+    color 120ms ease,
+    opacity 120ms ease,
+    border-color 120ms ease;
+
+  &[data-variant="drag"] {
+    cursor: grab;
+  }
+
+  &:hover {
+    background: ${({ theme }) =>
+      theme.scheme === "dark" ? "rgba(255, 255, 255, 0.08)" : "rgba(15, 23, 42, 0.06)"};
+    border-color: ${({ theme }) =>
+      theme.scheme === "dark" ? "rgba(148, 163, 184, 0.26)" : "rgba(71, 85, 105, 0.2)"};
+    color: var(--color-gray12);
+    opacity: 1;
+  }
+
+  &:focus-visible {
+    outline: 2px solid rgba(59, 130, 246, 0.5);
+    outline-offset: 1px;
+  }
+`
+
+export const BlockHandleGrip = styled.span`
+  display: grid;
+  grid-template-columns: repeat(2, 0.18rem);
+  grid-auto-rows: 0.18rem;
+  gap: 0.12rem;
+
+  span {
+    width: 0.18rem;
+    height: 0.18rem;
+    border-radius: 999px;
+    background: currentColor;
+    opacity: 0.78;
+  }
+`
+
+export const BlockHandlePlus = styled.span`
+  position: relative;
+  width: 0.82rem;
+  height: 0.82rem;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+
+  span {
+    position: absolute;
+    display: block;
+    border-radius: 999px;
+    background: currentColor;
+  }
+
+  span:first-of-type {
+    width: 0.82rem;
+    height: 1.6px;
+  }
+
+  span:last-of-type {
+    width: 1.6px;
+    height: 0.82rem;
+  }
+`
+
+export const DraggedBlockGhost = styled.div`
+  position: fixed;
+  z-index: 58;
+  pointer-events: none;
+  transform: translate3d(0, 0, 0);
+  filter: drop-shadow(0 20px 30px rgba(15, 23, 42, 0.3));
+`
+
+export const DraggedBlockGhostBadge = styled.div`
+  display: inline-flex;
+  align-items: center;
+  gap: 0.35rem;
+  min-height: 1.55rem;
+  margin: 0 0 0.32rem 0.18rem;
+  padding: 0 0.56rem;
+  border-radius: 999px;
+  border: 1px solid rgba(59, 130, 246, 0.34);
+  background: ${({ theme }) =>
+    theme.scheme === "dark" ? "rgba(17, 24, 39, 0.92)" : "rgba(248, 250, 252, 0.96)"};
+  color: ${({ theme }) => theme.colors.blue4};
+
+  span {
+    font-size: 0.72rem;
+    font-weight: 700;
+  }
+
+  strong {
+    font-size: 0.72rem;
+    font-weight: 700;
+    color: ${({ theme }) => theme.colors.blue3};
+  }
+`
+
+export const DraggedBlockGhostCard = styled.div`
+  overflow: hidden;
+  border-radius: 1rem;
+  border: 1px solid rgba(59, 130, 246, 0.28);
+  background: ${({ theme }) =>
+    theme.scheme === "dark" ? "rgba(15, 23, 42, 0.9)" : "rgba(255, 255, 255, 0.96)"};
+  box-shadow:
+    0 0 0 1px rgba(59, 130, 246, 0.18),
+    0 10px 22px rgba(15, 23, 42, 0.22);
+  padding: 0.72rem 0.88rem;
+  opacity: 0.92;
+  color: ${({ theme }) => theme.colors.gray12};
+  font-size: 0.82rem;
+  line-height: 1.45;
+  white-space: pre-wrap;
+  word-break: break-word;
+`
+
+export const BlockDropIndicator = styled.div`
+  position: fixed;
+  z-index: 56;
+  height: 4px;
+  border-radius: 999px;
+  background: linear-gradient(90deg, rgba(37, 99, 235, 0.95), rgba(59, 130, 246, 0.98));
+  box-shadow:
+    0 0 0 1px rgba(37, 99, 235, 0.2),
+    0 4px 10px rgba(37, 99, 235, 0.22);
+  pointer-events: none;
+
+  &::before,
+  &::after {
+    content: "";
+    position: absolute;
+    top: 50%;
+    width: 8px;
+    height: 8px;
+    border-radius: 999px;
+    background: rgba(59, 130, 246, 0.98);
+    box-shadow: 0 0 0 1px rgba(37, 99, 235, 0.28);
+    transform: translateY(-50%);
+  }
+
+  &::before {
+    left: -3px;
+  }
+
+  &::after {
+    right: -3px;
+  }
+`
+
+export const BlockSelectionOverlay = styled.div`
+  position: absolute;
+  z-index: 2;
+  pointer-events: none;
+  border-radius: 0.95rem;
+  background: ${({ theme }) =>
+    theme.scheme === "dark" ? "rgba(59, 130, 246, 0.12)" : "rgba(59, 130, 246, 0.1)"};
+  box-shadow: none;
+`
+
+export const MobileBlockActionBar = styled.div`
+  position: fixed;
+  z-index: 55;
+  display: flex;
+  flex-wrap: wrap;
+  gap: 0.45rem;
+  padding: 0.5rem;
+  border-radius: 0.9rem;
+  border: 1px solid ${({ theme }) => theme.colors.gray6};
+  background: ${({ theme }) =>
+    theme.scheme === "dark" ? "rgba(15, 18, 24, 0.96)" : "rgba(255, 255, 255, 0.98)"};
+  box-shadow: ${({ theme }) =>
+    theme.scheme === "dark" ? "0 10px 18px rgba(0, 0, 0, 0.14)" : "0 10px 18px rgba(15, 23, 42, 0.1)"};
+  transform: translateX(-50%);
+`
+
+export const FloatingBlockMenu = styled.div`
+  position: fixed;
+  z-index: 65;
+  width: min(30rem, calc(100vw - 2rem));
+  display: flex;
+  flex-direction: column;
+  gap: 0.6rem;
+  padding: 0.75rem;
+  border-radius: 0.9rem;
+  border: 1px solid ${({ theme }) => theme.colors.gray6};
+  background: ${({ theme }) =>
+    theme.scheme === "dark" ? "rgba(15, 18, 24, 0.96)" : "rgba(255, 255, 255, 0.98)"};
+  box-shadow: ${({ theme }) =>
+    theme.scheme === "dark" ? "0 14px 22px rgba(0, 0, 0, 0.15)" : "0 14px 22px rgba(15, 23, 42, 0.1)"};
+`
+
+export const FloatingBlockMenuHeader = styled.strong`
+  font-size: 0.88rem;
+  color: var(--color-gray12);
+`
+
+export const FloatingBlockMenuDivider = styled.div`
+  height: 1px;
+  background: ${({ theme }) =>
+    theme.scheme === "dark" ? "rgba(255, 255, 255, 0.06)" : "rgba(15, 23, 42, 0.08)"};
+`
+
+export const FloatingBlockMenuGrid = styled.div`
+  display: grid;
+  gap: 0.45rem;
+  grid-template-columns: repeat(auto-fit, minmax(9rem, 1fr));
+`
+
+export const FloatingBlockMenuButton = styled.button`
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 0.15rem;
+  min-height: 3rem;
+  border-radius: 0.85rem;
+  border: 1px solid ${({ theme }) => theme.colors.gray6};
+  background: ${({ theme }) =>
+    theme.scheme === "dark" ? "rgba(18, 21, 26, 0.72)" : "rgba(255, 255, 255, 0.96)"};
+  color: var(--color-gray12);
+  padding: 0.68rem 0.8rem;
+  text-align: left;
+
+  strong {
+    font-size: 0.82rem;
+  }
+
+  span {
+    color: var(--color-gray10);
+    font-size: 0.72rem;
+  }
+`
+
+export const FloatingBlockActionList = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 0.35rem;
+`
+
+export const FloatingBlockActionButton = styled.button`
+  min-height: 2.25rem;
+  border-radius: 0.8rem;
+  border: 1px solid ${({ theme }) => theme.colors.gray6};
+  background: ${({ theme }) =>
+    theme.scheme === "dark" ? "rgba(18, 21, 26, 0.72)" : "rgba(255, 255, 255, 0.96)"};
+  color: var(--color-gray12);
+  font-size: 0.8rem;
+  font-weight: 700;
+  text-align: left;
+  padding: 0 0.8rem;
+
+  &[data-active="true"] {
+    border-color: rgba(59, 130, 246, 0.48);
+    background: ${({ theme }) =>
+      theme.scheme === "dark" ? "rgba(37, 99, 235, 0.18)" : "rgba(219, 234, 254, 0.96)"};
+    color: ${({ theme }) => (theme.scheme === "dark" ? "#dbeafe" : "#1d4ed8")};
+  }
+
+  &[data-variant="danger"] {
+    border-color: rgba(248, 113, 113, 0.16);
+    color: #fecaca;
+    background: rgba(127, 29, 29, 0.1);
+  }
+`
+
+export const AuxDisclosure = styled.details`
+  border: 0;
+  border-top: 1px solid ${({ theme }) => theme.colors.gray6};
+  background: transparent;
+
+  > summary {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    gap: 0.75rem;
+    cursor: pointer;
+    list-style: none;
+    padding: 0.8rem 0;
+
+    &::-webkit-details-marker {
+      display: none;
+    }
+  }
+
+  strong {
+    font-size: 0.82rem;
+    color: var(--color-gray11);
+  }
+
+  span {
+    font-size: 0.76rem;
+    color: var(--color-gray10);
+  }
+
+  .body {
+    padding: 0 0 0.75rem;
+  }
+`
+
+export const QuickInsertBar = styled.div`
+  display: grid;
+  gap: 0.7rem;
+  padding: 0.1rem 0 0.2rem;
+`
+
+export const QuickInsertActions = styled.div`
+  display: flex;
+  flex-wrap: wrap;
+  min-width: 0;
+  max-width: 100%;
+  gap: 0.55rem;
+`
+
+export const QuickInsertButton = styled.button`
+  min-height: 2.4rem;
+  border-radius: 999px;
+  border: 1px solid ${({ theme }) =>
+    theme.scheme === "dark" ? "rgba(148, 163, 184, 0.22)" : "rgba(71, 85, 105, 0.14)"};
+  background: ${({ theme }) =>
+    theme.scheme === "dark" ? "rgba(17, 24, 39, 0.78)" : "rgba(255, 255, 255, 0.94)"};
+  color: var(--color-gray12);
+  font-size: 0.82rem;
+  font-weight: 700;
+  white-space: nowrap;
+  padding: 0 0.95rem;
+  transition:
+    transform 120ms ease,
+    border-color 120ms ease,
+    background 120ms ease;
+
+  &:hover:not(:disabled),
+  &:focus-visible:not(:disabled) {
+    border-color: ${({ theme }) => theme.colors.blue7};
+    transform: translateY(-1px);
+  }
+
+  &:disabled {
+    cursor: not-allowed;
+    opacity: 0.5;
+  }
+`
