+++ front/src/components/editor/BlockEditorEngine.toolbarStyles.tsx
@@ -1,0 +1,590 @@
+import { keyframes } from "@emotion/react"
+import styled from "@emotion/styled"
+
+export const Shell = styled.div`
+  display: flex;
+  flex-direction: column;
+  min-width: 0;
+  max-width: 100%;
+  gap: 0.8rem;
+`
+
+export const Toolbar = styled.div`
+  display: flex;
+  flex-wrap: wrap;
+  align-items: center;
+  justify-content: space-between;
+  width: 100%;
+  max-width: 100%;
+  min-width: 0;
+  gap: 0.9rem;
+  padding: 0 0 0.7rem;
+  overflow-x: clip;
+  border-bottom: 1px solid rgba(148, 163, 184, 0.14);
+  background: transparent;
+`
+
+export const ToolbarActions = styled.div`
+  display: inline-flex;
+  flex: 1 1 100%;
+  flex-wrap: wrap;
+  align-items: center;
+  width: 100%;
+  max-width: 100%;
+  gap: 0.25rem;
+  min-width: 0;
+`
+
+export const ToolbarGroup = styled.div`
+  display: inline-flex;
+  align-items: center;
+  gap: 0.14rem;
+`
+
+export const ToolbarSeparator = styled.span`
+  width: 1px;
+  height: 1.7rem;
+  background: ${({ theme }) =>
+    theme.scheme === "dark" ? "rgba(148, 163, 184, 0.18)" : "rgba(15, 23, 42, 0.12)"};
+`
+
+export const ToolbarMoreDisclosure = styled.details`
+  position: relative;
+  display: inline-flex;
+  flex-direction: column;
+
+  summary {
+    list-style: none;
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    width: 2.4rem;
+    height: 2.4rem;
+    border-radius: 0.8rem;
+    border: 0;
+    background: transparent;
+    color: ${({ theme }) => theme.colors.gray11};
+    font-size: 1.3rem;
+    font-weight: 700;
+    cursor: pointer;
+    transition: background-color 160ms ease, color 160ms ease;
+  }
+
+  summary:hover {
+    background: ${({ theme }) =>
+      theme.scheme === "dark" ? "rgba(148, 163, 184, 0.08)" : "rgba(15, 23, 42, 0.05)"};
+    color: var(--color-gray12);
+  }
+
+  summary::-webkit-details-marker {
+    display: none;
+  }
+
+  .body {
+    position: absolute;
+    top: calc(100% + 0.55rem);
+    right: 0;
+    z-index: 30;
+    display: flex;
+    flex-wrap: wrap;
+    gap: 0.55rem;
+    min-width: 16rem;
+    padding: 0.8rem;
+    border-radius: 1rem;
+    border: 1px solid ${({ theme }) => theme.colors.gray6};
+    background: ${({ theme }) =>
+      theme.scheme === "dark" ? "rgba(12, 16, 22, 0.96)" : "rgba(255, 255, 255, 0.98)"};
+    box-shadow: ${({ theme }) =>
+      theme.scheme === "dark"
+        ? "0 18px 40px rgba(3, 7, 18, 0.32)"
+        : "0 18px 40px rgba(15, 23, 42, 0.12)"};
+  }
+`
+
+export const ToolbarColorDisclosure = styled.details`
+  position: relative;
+  display: inline-flex;
+  flex-direction: column;
+
+  summary {
+    list-style: none;
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    min-width: 2.7rem;
+    height: 2.4rem;
+    padding: 0 0.45rem;
+    border-radius: 0.8rem;
+    border: 0;
+    background: transparent;
+    color: ${({ theme }) => theme.colors.gray11};
+    cursor: pointer;
+    transition: background-color 160ms ease, color 160ms ease;
+  }
+
+  summary[data-active="true"] {
+    background: ${({ theme }) =>
+      theme.scheme === "dark" ? "rgba(148, 163, 184, 0.14)" : "rgba(15, 23, 42, 0.08)"};
+    color: ${({ theme }) => theme.colors.gray12};
+    box-shadow: inset 0 -1.5px 0
+      ${({ theme }) =>
+        theme.scheme === "dark" ? "rgba(226, 232, 240, 0.32)" : "rgba(15, 23, 42, 0.22)"};
+  }
+
+  summary:hover {
+    background: ${({ theme }) =>
+      theme.scheme === "dark" ? "rgba(148, 163, 184, 0.08)" : "rgba(15, 23, 42, 0.05)"};
+    color: var(--color-gray12);
+  }
+
+  summary::-webkit-details-marker {
+    display: none;
+  }
+
+  .body {
+    position: absolute;
+    top: calc(100% + 0.55rem);
+    right: 0;
+    z-index: 32;
+    display: grid;
+    gap: 0.42rem;
+    min-width: 10.5rem;
+    padding: 0.72rem;
+    border-radius: 1rem;
+    border: 1px solid ${({ theme }) => theme.colors.gray6};
+    background: ${({ theme }) =>
+      theme.scheme === "dark" ? "rgba(12, 16, 22, 0.96)" : "rgba(255, 255, 255, 0.98)"};
+    box-shadow: ${({ theme }) =>
+      theme.scheme === "dark"
+        ? "0 18px 40px rgba(3, 7, 18, 0.32)"
+        : "0 18px 40px rgba(15, 23, 42, 0.12)"};
+  }
+`
+
+export const ColorTriggerIcon = styled.span`
+  display: inline-flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 0.1rem;
+  min-width: 1.15rem;
+
+  span {
+    font-size: 0.96rem;
+    font-weight: 760;
+    line-height: 1;
+    letter-spacing: -0.02em;
+  }
+
+  i {
+    display: block;
+    width: 1rem;
+    height: 0.18rem;
+    border-radius: 999px;
+    background: ${({ theme }) => theme.colors.gray8};
+  }
+
+  &[data-active="true"] i {
+    box-shadow: 0 0 0 1px rgba(255, 255, 255, 0.08);
+  }
+`
+
+export const ColorOptionButton = styled.button`
+  min-height: 2rem;
+  border-radius: 0.8rem;
+  border: 1px solid ${({ theme }) => theme.colors.gray6};
+  background: ${({ theme }) =>
+    theme.scheme === "dark" ? "rgba(18, 21, 26, 0.42)" : "rgba(255, 255, 255, 0.96)"};
+  color: var(--color-gray12);
+  padding: 0 0.72rem;
+  text-align: left;
+
+  &[data-active="true"] {
+    border-color: ${({ theme }) =>
+      theme.scheme === "dark" ? "rgba(59, 130, 246, 0.32)" : "rgba(37, 99, 235, 0.24)"};
+    background: ${({ theme }) =>
+      theme.scheme === "dark" ? "rgba(37, 99, 235, 0.12)" : "rgba(37, 99, 235, 0.08)"};
+  }
+
+  &:disabled {
+    opacity: 0.44;
+    cursor: not-allowed;
+  }
+`
+
+export const ColorOptionLabel = styled.span`
+  display: inline-flex;
+  align-items: center;
+  gap: 0.5rem;
+  font-size: 0.78rem;
+  font-weight: 700;
+`
+
+export const ColorOptionSwatch = styled.span`
+  display: inline-flex;
+  width: 0.92rem;
+  height: 0.92rem;
+  border-radius: 999px;
+  border: 1px solid ${({ theme }) => theme.colors.gray6};
+  background: ${({ theme }) => theme.colors.gray3};
+
+  &[data-empty="true"] {
+    position: relative;
+    background: transparent;
+  }
+
+  &[data-empty="true"]::after {
+    content: "";
+    position: absolute;
+    inset: 0.38rem -0.05rem auto -0.05rem;
+    height: 1.5px;
+    background: ${({ theme }) => theme.colors.gray10};
+    transform: rotate(-34deg);
+    transform-origin: center;
+  }
+`
+
+export const ToolbarRibbonButton = styled.button`
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  min-width: 2.5rem;
+  height: 2.4rem;
+  padding: 0 0.58rem;
+  border: 0;
+  border-radius: 0.8rem;
+  background: transparent;
+  color: ${({ theme }) => theme.colors.gray11};
+  font-size: 1.02rem;
+  font-weight: 700;
+  line-height: 1;
+  letter-spacing: -0.02em;
+  transition: background-color 160ms ease, color 160ms ease;
+
+  svg {
+    width: 1.2rem;
+    height: 1.2rem;
+  }
+
+  &[data-tone="heading"] {
+    min-width: 3.1rem;
+    color: ${({ theme }) => theme.colors.gray10};
+    font-family: Georgia, "Times New Roman", serif;
+    font-size: 1rem;
+    font-weight: 700;
+  }
+
+  &[data-active="true"] {
+    background: ${({ theme }) =>
+      theme.scheme === "dark" ? "rgba(148, 163, 184, 0.14)" : "rgba(15, 23, 42, 0.08)"};
+    color: ${({ theme }) => theme.colors.gray12};
+    box-shadow: inset 0 -1.5px 0
+      ${({ theme }) =>
+        theme.scheme === "dark" ? "rgba(226, 232, 240, 0.32)" : "rgba(15, 23, 42, 0.22)"};
+  }
+
+  &:hover {
+    background: ${({ theme }) =>
+      theme.scheme === "dark" ? "rgba(148, 163, 184, 0.08)" : "rgba(15, 23, 42, 0.05)"};
+    color: var(--color-gray12);
+  }
+
+  &:disabled {
+    opacity: 0.44;
+    cursor: not-allowed;
+  }
+`
+
+export const ToolbarButton = styled.button`
+  min-height: 2rem;
+  border-radius: 999px;
+  border: 1px solid ${({ theme }) => theme.colors.gray6};
+  background: ${({ theme }) =>
+    theme.scheme === "dark" ? "rgba(18, 21, 26, 0.42)" : "rgba(255, 255, 255, 0.96)"};
+  color: var(--color-gray11);
+  font-size: 0.78rem;
+  font-weight: 700;
+  padding: 0 0.82rem;
+
+  &[data-active="true"] {
+    border-color: ${({ theme }) =>
+      theme.scheme === "dark" ? "rgba(59, 130, 246, 0.32)" : "rgba(37, 99, 235, 0.24)"};
+    background: ${({ theme }) =>
+      theme.scheme === "dark" ? "rgba(37, 99, 235, 0.12)" : "rgba(37, 99, 235, 0.08)"};
+    color: ${({ theme }) => (theme.scheme === "dark" ? "#bfdbfe" : theme.colors.blue8)};
+  }
+
+  &[data-variant="subtle-danger"] {
+    border-color: rgba(248, 113, 113, 0.14);
+    color: #fda4af;
+  }
+
+  &[data-variant="danger"] {
+    border-color: rgba(248, 113, 113, 0.22);
+    background: rgba(127, 29, 29, 0.12);
+    color: #fecdd3;
+  }
+
+  &:disabled {
+    opacity: 0.48;
+    cursor: not-allowed;
+  }
+`
+
+export const HiddenFileInput = styled.input`
+  display: none;
+`
+
+export const TablePresetSwatches = styled.div`
+  display: inline-flex;
+  align-items: center;
+  gap: 0.28rem;
+`
+
+export const TablePresetSwatch = styled.button`
+  --table-swatch-color: #dbeafe;
+  width: 1.5rem;
+  height: 1.5rem;
+  border-radius: 999px;
+  border: 1px solid ${({ theme }) => theme.colors.gray6};
+  background: var(--table-swatch-color);
+  padding: 0;
+
+  &[data-active="true"] {
+    box-shadow: 0 0 0 2px rgba(59, 130, 246, 0.28);
+    border-color: rgba(59, 130, 246, 0.42);
+  }
+`
+
+export const TableColorInput = styled.input`
+  width: 2.2rem;
+  height: 2rem;
+  border: 1px solid ${({ theme }) => theme.colors.gray6};
+  border-radius: 999px;
+  background: transparent;
+  padding: 0.22rem;
+`
+
+export const slashMenuFadeInFromBottom = keyframes`
+  from {
+    opacity: 0;
+    transform: translateY(6px) scale(0.985);
+  }
+
+  to {
+    opacity: 1;
+    transform: translateY(0) scale(1);
+  }
+`
+
+export const slashMenuFadeInFromTop = keyframes`
+  from {
+    opacity: 0;
+    transform: translateY(-6px) scale(0.985);
+  }
+
+  to {
+    opacity: 1;
+    transform: translateY(0) scale(1);
+  }
+`
+
+export const SlashMenu = styled.div`
+  position: fixed;
+  z-index: 70;
+  display: flex;
+  flex-direction: column;
+  gap: 0.4rem;
+  width: min(38rem, calc(100vw - 1.5rem));
+  padding: 0.45rem 0.45rem 0.35rem;
+  border: 1px solid ${({ theme }) => theme.colors.gray6};
+  border-radius: 1rem;
+  background: ${({ theme }) =>
+    theme.scheme === "dark" ? "rgba(28, 28, 28, 0.98)" : "rgba(255, 255, 255, 0.98)"};
+  box-shadow: ${({ theme }) =>
+    theme.scheme === "dark" ? "0 18px 36px rgba(0, 0, 0, 0.22)" : "0 18px 36px rgba(15, 23, 42, 0.14)"};
+  backdrop-filter: blur(12px);
+  transform-origin: top left;
+  animation: ${slashMenuFadeInFromBottom} 140ms cubic-bezier(0.2, 0.8, 0.2, 1);
+  transition: left 120ms cubic-bezier(0.2, 0.8, 0.2, 1), top 120ms cubic-bezier(0.2, 0.8, 0.2, 1),
+    box-shadow 140ms ease, border-color 140ms ease;
+  will-change: left, top, transform, opacity;
+
+  &[data-placement="top"] {
+    transform-origin: bottom left;
+    animation-name: ${slashMenuFadeInFromTop};
+  }
+
+  @media (prefers-reduced-motion: reduce) {
+    animation: none;
+  }
+`
+
+export const SlashQuerySummary = styled.div`
+  display: inline-flex;
+  align-items: center;
+  min-height: 2.35rem;
+  border-radius: 0.75rem;
+  border: 0;
+  background: ${({ theme }) =>
+    theme.scheme === "dark" ? "rgba(255, 255, 255, 0.04)" : "rgba(15, 23, 42, 0.04)"};
+  padding: 0 0.75rem;
+
+  span {
+    color: var(--color-gray10);
+    font-size: 0.92rem;
+    font-weight: 600;
+  }
+`
+
+export const SlashMenuBody = styled.div`
+  display: grid;
+  gap: 0.4rem;
+  max-height: min(62vh, 30rem);
+  overflow-y: auto;
+  padding: 0.1rem 0.1rem 0.15rem 0;
+
+  scrollbar-width: thin;
+
+  &::-webkit-scrollbar {
+    width: 8px;
+  }
+
+  &::-webkit-scrollbar-thumb {
+    border-radius: 999px;
+    background: ${({ theme }) =>
+      theme.scheme === "dark" ? "rgba(148, 163, 184, 0.28)" : "rgba(100, 116, 139, 0.24)"};
+  }
+`
+
+export const SlashMenuSection = styled.section`
+  display: grid;
+  gap: 0.15rem;
+
+  & + & {
+    margin-top: 0.15rem;
+    padding-top: 0.55rem;
+    border-top: 1px solid ${({ theme }) =>
+      theme.scheme === "dark" ? "rgba(255, 255, 255, 0.04)" : "rgba(15, 23, 42, 0.05)"};
+  }
+`
+
+export const SlashMenuSectionLabel = styled.strong`
+  display: inline-flex;
+  align-items: center;
+  min-height: 1.2rem;
+  padding: 0 0.5rem;
+  color: var(--color-gray10);
+  font-size: 0.74rem;
+  font-weight: 800;
+`
+
+export const SlashActionList = styled.div`
+  display: grid;
+  gap: 0.3rem;
+`
+
+export const SlashActionIcon = styled.span`
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  min-width: 1.8rem;
+  height: 1.8rem;
+  border-radius: 0;
+  border: 0;
+  background: transparent;
+  color: var(--color-gray12);
+  font-size: 0.92rem;
+  font-weight: 700;
+  letter-spacing: -0.02em;
+`
+
+export const SlashActionMain = styled.span`
+  display: grid;
+  gap: 0.02rem;
+  text-align: left;
+
+  strong {
+    font-size: 0.88rem;
+    color: var(--color-gray12);
+    font-weight: 700;
+  }
+
+  span {
+    color: var(--color-gray10);
+    font-size: 0.75rem;
+    line-height: 1.35;
+  }
+`
+
+export const SlashActionTitleRow = styled.span`
+  display: inline-flex;
+  align-items: center;
+  gap: 0.25rem;
+  flex-wrap: wrap;
+`
+
+export const SlashActionHint = styled.span`
+  color: var(--color-gray10);
+  font-size: 0.76rem;
+  font-weight: 700;
+  letter-spacing: -0.01em;
+`
+
+export const SlashActionButton = styled.button`
+  display: grid;
+  grid-template-columns: auto minmax(0, 1fr) auto;
+  align-items: center;
+  gap: 0.55rem;
+  min-height: 2.7rem;
+  border-radius: 0.75rem;
+  border: 0;
+  background: transparent;
+  padding: 0.45rem 0.55rem;
+  text-align: left;
+  transition: background-color 120ms ease, color 120ms ease;
+
+  &:hover:not(:disabled),
+  &:focus-visible:not(:disabled) {
+    background: ${({ theme }) =>
+      theme.scheme === "dark" ? "rgba(255, 255, 255, 0.05)" : "rgba(15, 23, 42, 0.04)"};
+  }
+
+  &[data-active="true"] {
+    background: ${({ theme }) =>
+      theme.scheme === "dark" ? "rgba(255, 255, 255, 0.07)" : "rgba(15, 23, 42, 0.06)"};
+  }
+
+  &[data-active="true"][data-input-mode="keyboard"] {
+    background: ${({ theme }) =>
+      theme.scheme === "dark" ? "rgba(255, 255, 255, 0.08)" : "rgba(15, 23, 42, 0.07)"};
+  }
+
+  &[data-active="true"][data-input-mode="pointer"] {
+    background: ${({ theme }) =>
+      theme.scheme === "dark" ? "rgba(255, 255, 255, 0.07)" : "rgba(15, 23, 42, 0.06)"};
+  }
+
+  &[data-active="true"] [data-role="slash-action-icon"] {
+    color: ${({ theme }) => theme.colors.gray12};
+  }
+
+  &:disabled {
+    opacity: 0.45;
+    cursor: not-allowed;
+  }
+
+  @media (prefers-reduced-motion: reduce) {
+    transition: none;
+  }
+`
+
+export const SlashEmptyState = styled.div`
+  display: grid;
+  place-items: center;
+  min-height: 6rem;
+  border-radius: 0.75rem;
+  border: 0;
+  color: var(--color-gray10);
+  font-size: 0.8rem;
+  font-weight: 600;
+`
